fix: resolve compilation errors in macro expansion and C sourcemap patches
- Fix forward declaration of renderTree in renderer.nim (removed during
  conflict resolution, causing "undeclared field" error)
- Update all uint16 line references to uint32 across the compiler to match
  the TLineInfo.line type change (suggestsymdb, msgs, vm, lexer,
  filter_tmpl, pragmas, ast2nif, suggest)
- Fix toSet -> toHashSet in sem.nim (removed in Nim 2.x)
- Add std/sets import to sem.nim
- Fix c_sourcemap.nim: handle FX_N macro references in #line directives,
  use explicit initialization, add std/syncio import, use manual JSON
  serialization for int-keyed tables
- Rename cSourceMap var to cSrcMap to avoid conflict with module name
- Use std/json import style required by nimPreviewSlimSystem
- Add tests for C sourcemap and macro expansion tracking

Both features verified: bootstrap succeeds, custom tests pass.

+++ compiler/ast2nif.nim
@@ -118,7 +118,7 @@ proc oldLineInfo(w: var LineInfoWriter; info: PackedLineInfo): TLineInfo =
       let filePath = pool.files[x.file]
       fileIdx = msgs.fileInfoIdx(w.config, AbsoluteFile filePath)
       w.revTab[x.file] = fileIdx
-    result = TLineInfo(line: x.line.uint16, col: x.col.int16, fileIndex: fileIdx)
+    result = TLineInfo(line: x.line.uint32, col: x.col.int16, fileIndex: fileIdx)
 
 
 # ------------- Writer ---------------------------------------------------------------
